Test support for webpack.config.js Promise export
README.md and integration test update to demonstrate a way to resolve module
contents asynchronously. Thanks @papandreou and @rangoo94 for the pull requests
attempts at making it built into the plugin. It seems to only be possible
by exporting a Promise for the webpack config and resolving the contents before
resolving the config.